+++ config/config.exs
@@ -3,7 +3,6 @@
 use Mix.Config
 
 config :bolt,
-  base_doc_url: System.get_env("BASE_DOC_URL"),
   botlog_channel: System.get_env("BOTLOG_CHANNEL"),
   ecto_repos: [Bolt.Repo],
   prefix: System.get_env("BOT_PREFIX") || ".",
@@ -13,7 +12,8 @@ config :bolt,
     |> Enum.map(fn user_id ->
       {value, _} = Integer.parse(user_id)
       value
-    end)
+    end),
+  web_domain: System.get_env("WEB_DOMAIN")
 
 config :nosedrum,
   prefix: System.get_env("BOT_PREFIX") || "."
@@ -30,6 +30,7 @@ config :crow,
   ip: {127, 0, 0, 1},
   plugins: [
     Bolt.CrowPlugins.GuildMembers,
+    Bolt.CrowPlugins.GuildMessageCounts,
     CrowPlugins.BEAM.ContextSwitches,
     CrowPlugins.BEAM.GarbageCollections,
     CrowPlugins.BEAM.IO,