+++ lib/bolt/cogs/help.ex
@@ -60,7 +60,7 @@ defmodule Bolt.Cogs.Help do
       footer: %Embed.Footer{
         text:
           "Online documentation: " <>
-            "#{Application.get_env(:bolt, :base_doc_url) || "(not set on this instance)"}"
+            "https://#{Application.get_env(:bolt, :web_domain) || "(not set on this instance)"}"
       },
       color: Constants.color_blue()
     }