+++ lib/bolt/consumer/message_create.ex
@@ -3,6 +3,7 @@ defmodule Bolt.Consumer.MessageCreate do
 
   @nosedrum_storage_implementation Nosedrum.Storage.ETS
 
+  alias Bolt.CrowPlugins.GuildMessageCounts
   alias Bolt.{MessageCache, USW}
   alias Nosedrum.Invoker.Split, as: CommandInvoker
   alias Nostrum.Struct.Message
@@ -13,6 +14,7 @@ defmodule Bolt.Consumer.MessageCreate do
       CommandInvoker.handle_message(msg, @nosedrum_storage_implementation)
       MessageCache.consume(msg)
       USW.apply(msg)
+      GuildMessageCounts.record_message(msg)
     end
   end
 end